+++ vector/src/main/java/im/vector/riotx/features/crypto/verification/VerificationBottomSheetViewModel.kt
@@ -128,9 +128,11 @@ class VerificationBottomSheetViewModel @AssistedInject constructor(@Assisted ini
                         )
                     }
                     val roomParams = CreateRoomParams(
-                            invitedUserIds = listOf(otherUserId).toMutableList()
+                            invitedUserIds = listOf(otherUserId)
                     )
                             .setDirectMessage()
+                            .also { it.enableEncryptionIfInvitedUsersSupportIt = true }
+
                     session.createRoom(roomParams, object : MatrixCallback<String> {
                         override fun onSuccess(data: String) {
                             setState {